+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Application/SettlementReports/ISettlementReportRepository.cs
@@ -12,12 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Globalization;
-using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
+namespace Energinet.DataHub.Wholesale.CalculationResults.Application.SettlementReports;
 
-namespace Energinet.DataHub.Wholesale.CalculationResults.Application;
-
-public interface ISettlementReportResultsCsvWriter
+public interface ISettlementReportRepository
 {
-    Task WriteAsync(Stream destination, IEnumerable<SettlementReportResultRow> rows, CultureInfo csvFormatLocale);
+    /// <summary>
+    /// Create zip archives for each process in the batch.
+    /// The archive contains the basis data files and the result file.
+    /// </summary>
+    Task CreateSettlementReportsAsync(BatchInfo completedBatchInfo);
+
+    Task<SettlementReport> GetSettlementReportAsync(BatchInfo batchInfo);
+
+    Task GetSettlementReportAsync(BatchInfo completedBatchInfo, string gridAreaCode, Stream outputStream);
 }
