+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Application/SettlementReports/SettlementReport.cs
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.CalculationResults.Interfaces.SettlementReport;
+namespace Energinet.DataHub.Wholesale.CalculationResults.Application.SettlementReports;
 
 public record SettlementReport(Stream Stream);
