+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/SettlementReports/SettlementReportResultsCsvWriter.cs
@@ -16,6 +16,7 @@ using System.Globalization;
 using CsvHelper;
 using CsvHelper.Configuration;
 using Energinet.DataHub.Wholesale.CalculationResults.Application;
+using Energinet.DataHub.Wholesale.CalculationResults.Application.SettlementReports;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
 using Energinet.DataHub.Wholesale.Common.Models;
